+++ python/cusignal/_upfirdn.py
@@ -19,8 +19,14 @@ import warnings
 from enum import Enum
 from math import ceil
 from numba import complex64, complex128, cuda, float32, float64, int64, void
-from numba.types.scalars import Complex
 from string import Template
+
+try:
+    # Numba <= 0.49
+    from numba.types.scalars import Complex
+except ImportError:
+    # Numba >= 0.49
+    from numba.core.types.scalars import Complex
 
 
 class GPUKernel(Enum):