+++ python/cusignal/signaltools.py
@@ -1808,7 +1808,8 @@ def decimate(x, q, n=None, axis=-1, zero_phase=True):
         if n is None:
             half_len = 10 * q  # reasonable cutoff for our sinc-like function
             n = 2 * half_len
-        b = firwin(n + 1, 1.0 / q, window="hamming")
+
+        b = firwin(n + 1, 1.0 / q, window='hamming')
 
     sl = [slice(None)] * x.ndim